tests and fixes for MultiPhraseQuery, toArray() must set now whether or not to null terminate, various unsigned compare fixes,

+++ src/core/CLucene/search/BooleanQuery.cpp
@@ -335,7 +335,7 @@ CL_NS_DEF(search)
                                                  parentQuery->minNrShouldMatch,
                                                  parentQuery->allowDocsOutOfOrder);
 
-      for (int i = 0 ; i < weights.size(); i++) {
+      for (size_t i = 0 ; i < weights.size(); i++) {
         BooleanClause* c = (*clauses)[i];
         Weight* w = weights[i];
         Scorer* subScorer = w->scorer(reader);
